Create Verify Tag on Allowed Branch workflow
A workflow that will look at the github.ref along with
github.ref_name and github.ref_type to verify that
this workflow is triggered by a "on: tag:" event and
that the tag actually exists on a list of permitted
branch names.

The primary purpose is to prevent a version tag from
being pushed to the wrong branch.  You can't do that
with the workflow triggers, because when you specify
both "on: tags:" and "on: branches:" it will operate as
a logical "or".