fix: job steps which need to run in sub-directory
In situations where the job specifies a default
working-directory for a run statement, where that
directory only exists after checkout; we need to
reorder the steps.

The alternative choice would be to specify a "./"
working-directory for the steps that don't have to
be run from a subdirectory.  This is good for those
debug statements examining the github.ref value
prior to doing the checkout.