Spit out the report file in the last step
This tripped me up.  When the build failed, I went and
looked for the failed step and ended up here.  But it
confused me because there were no messages to
indicate why the build failed.  The actual failure
message was back up in the 'audit' step.

So to make things simple for those of us who haven't
had enough coffee in the morning, we'll spit out the
report text in this step for ease of use.