Add dotnet-npm-astro-build-test-publish.yml
A combined .NET / NPM / Astro build process that produces
an artifact for deployment.  It's useful in situations where you
are using Astro and the build directory is very large and/or
you don't want to persist between steps.

The downside, of course, is that there's more to dig through
in the log if this combination job fails.