Improved dotnet-publish.yml (BREAKING)
Adds a new required input to specify the publish
working directory instead of defaulting to the
project_directory.

This reflects changes that we had to do in the
JFrog-specific publish workflow in order to support
JFrog build-info JSON.

Since we need that change there, making it here
keeps the two versions of the publish workflow
more compatible with each other.