Add deploy-zip-to-azure-webapps-with-swap.yml
This version of the deploy step is designed for Linux Web App
instances in Azure which do not support the "auto-swap
when healthy" for the staging slot.

Deploy your artifact to the 'staging' slot as normal, then
use this action to wait for it to go health before swapping
the artifact into the 'production' slot.