Add actions/health-and-version-check@v1.17.0 steps
Add steps to the "with swap" workflow to call the new
actions/health-and-version-check GitHub action to validate
that the specified version string appears at the version URI.
If we don't see the version string, we should reboot once
and check again for another 300 seconds.

There is also a "with-verify" workflow added which can be
used for situations where we deploy to a specific slot and
won't be doing a "swap when ready" to the production slot.