Revert "Pin to v4.4.3 for jfrog/setup-jfrog-cli action"
This reverts commit e9b22c1ef1814756f10d335e870363a5e8f88c2b.

JFrog reversed on making the existing default
server ID different from what was there before
in version 4.5.1.  So now workflows using the v4
of the jfrog/setup-jfrog-cli  GitHub action will no
longer misbehave when they pull the latest for
the 'v4' tag.

Note that they still added the 'custom-server-id'
input so that we can start customizing it for our
needs and protect against future breakage. I'm
not sure if they will be reintroducing the change
in v5, so we might want to start specifying our
own custom server ID.