Pin to v4.5.6 of jfrog/setup-jfrog-cli for OIDC bug
v4.5.7 and later of the jfrog/setup-jfrog-cli GitHub Action
have broken the contract of what the default "audience"
OIDC value is being passed over to the JFrog Cloud instance
from the GitHub Actions run.

The default for this action used to be the GitHub organization's URL,
unless you passed in a value for 'oidc-audience' when calling the
setup action.  In v4.5.7, they changed the audience default to be
'jfrog-github' -- which breaks existing OIDC Integrations with GitHub.

The workaround for now is either to:

1. Pin to v4.5.6 instead of using the 'v4' tag.

2. Change all your workflows to pass the correct OIDC Audience value.

3. Change your OIDC Integration in the JFrog administration portal.

Based on conversations in issue 270 at https://github.com/jfrog/setup-jfrog-cli/issues
we believe that the change to the default value should be reverted and
we can eventually go back to using the 'v4' tag here.  We're trying to defer
making changes to our OIDC configuration until the outcome of issue
270 is decided.