Make rimbot_registry_access_key optional, add notes
This input is only used in some of our Terraform plan/apply cycles
where we deploy Azure Web Apps.  RIMbot is our internal GitHub
issue/PR tool that we use for deployments, restarts, swaps, etc.

In order for it to work, we have to keep a list of the web apps that
RIMbot is supposed to know about and we refresh that using our
Terraform plans and a Powershell script.